Changes in customer controllers

+++ StudioUp.Repo/Repositories/CustomerRepository.cs
@@ -23,18 +23,19 @@ namespace StudioUp.Repo.Repositories
             this.mapper = mapper;
             
         }
-        public async Task<int> AddAsync(CustomerDTO entity)
+        public async Task<DTO.CustomerDTO> AddAsync(CustomerDTO entity)
         {
             try
             {
                 var mapCast = mapper.Map<Models.Customer>(entity);
                 var newCustomer = await context.Customers.AddAsync(mapCast);
                 await context.SaveChangesAsync();
-                return newCustomer.Entity.Id;
+                entity.Id = newCustomer.Entity.Id;
+                return entity;
             }
             catch (Exception ex)
             {
-                return 0;
+                throw new Exception("Cann't add this object");
             }
         }
 
@@ -43,6 +44,10 @@ namespace StudioUp.Repo.Repositories
             try
             {
                 var c = await context.Customers.FirstOrDefaultAsync(t => t.Id == id);
+                if (c == null)
+                {
+                    return false;
+                }
                 var mapC = mapper.Map<Customer>(c);
                 context.Customers.Remove(mapC);
                 await context.SaveChangesAsync();
@@ -50,7 +55,8 @@ namespace StudioUp.Repo.Repositories
             }
             catch (Exception ex)
             {
-                return false;
+                throw new Exception("Cann't delete this object");
+
             }
         }
 
@@ -65,7 +71,8 @@ namespace StudioUp.Repo.Repositories
             }
             catch (Exception ex)
             {
-                return new List<CustomerDTO>();
+                throw new Exception("Not secceed");
+
             }
 
         }
@@ -75,13 +82,14 @@ namespace StudioUp.Repo.Repositories
             try
             {
                 var c = await context.Customers.FirstOrDefaultAsync(t => t.Id == id);
+
                 var mapCust = mapper.Map<CustomerDTO>(c);
                 return mapCust;
 
             }
             catch (Exception ex)
             {
-                return new CustomerDTO();
+                throw new Exception("Not secceed");
             }
         }
 
@@ -102,7 +110,7 @@ namespace StudioUp.Repo.Repositories
                 customerToUpdate.PaymentOptionId = entity.PaymentOptionId;
                 customerToUpdate.HMOId = entity.HMOId;
                 customerToUpdate.CustomerTypeId = entity.CustomerTypeId;
-                //customerToUpdate.IsActive = entity.IsActive;
+                customerToUpdate.IsActive = entity.IsActive;
                 customerToUpdate.SubscriptionTypeId = entity.SubscriptionTypeId;
                 customerToUpdate.Tel = entity.Tel;
                 context.Customers.Update(mapper.Map<Customer>(customerToUpdate));
@@ -113,7 +121,7 @@ namespace StudioUp.Repo.Repositories
             }
             catch (Exception ex)
             {
-                return false;
+                throw new Exception("Cann't update this object");
             }
         }
     }
